Fix return type for getMessageRegisterAddress func
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/sbe/+/176155
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>

+++ public/src/common/utils/sbeutil.C
@@ -37,7 +37,7 @@ namespace SBE
     bool isSimics() __attribute__((alias("__isSimicsRunning")));
     extern "C" void __isSimicsRunning() __attribute__ ((noinline));
 
-    uint64_t __attribute__((weak)) getMessageRegisterAddress()
+    uint32_t __attribute__((weak)) getMessageRegisterAddress()
     {
         SBE_DEBUG("getMessageRegisterAddress - default");
         return scomt::poz::FSXCOMP_FSXLOG_SB_MSG;
